aug25/using_llms_directly/pyproject.toml
[project]
name = "using-llms-directly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
]
